app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
yt-dlp>=2025.9.23
jinja2
sqlalchemy
aiosqlite
greenlet